+++ trans/views.py
@@ -1,4 +1,6 @@
 from django.shortcuts import render_to_response, get_object_or_404
+from django.views.decorators.cache import cache_page
+from django.conf import settings
 from django.core.servers.basehttp import FileWrapper
 from django.utils.translation import ugettext_lazy, ugettext as _
 from django.template import RequestContext, loader
@@ -16,6 +18,9 @@ from accounts.models import Profile
 from whoosh.analysis import StandardAnalyzer, StemmingAnalyzer
 import logging
 import os.path
+import json
+from xml.etree import ElementTree
+import urllib2
 
 # See https://code.djangoproject.com/ticket/6027
 class FixedFileWrapper(FileWrapper):
@@ -462,3 +467,36 @@ def not_found(request):
         'title': _('Page Not Found'),
         'projects': Project.objects.all(),
     })))
+
+# Cache this page for one day, it should not really change much
+@cache_page(24 * 3600)
+def js_config(request):
+    if settings.MT_APERTIUM_KEY is not None and settings.MT_APERTIUM_KEY != '':
+        try:
+            listpairs = urllib2.urlopen('http://api.apertium.org/json/listPairs?key=%s' % settings.MT_APERTIUM_KEY)
+            pairs = listpairs.read()
+            parsed = json.loads(pairs)
+            apertium_langs = [p['targetLanguage'] for p in parsed['responseData'] if p['sourceLanguage'] == 'en']
+        except Exception, e:
+            logger.error('failed to get supported languages from Apertium, using defaults (%s)', str(e))
+            apertium_langs = ['gl', 'ca', 'es', 'eo']
+    else:
+        apertium_langs = None
+
+    if settings.MT_MICROSOFT_KEY is not None and settings.MT_MICROSOFT_KEY != '':
+        try:
+            listpairs = urllib2.urlopen('http://api.microsofttranslator.com/V2/Http.svc/GetLanguagesForTranslate?appID=%s' % settings.MT_MICROSOFT_KEY)
+            data = listpairs.read()
+            parsed = ElementTree.fromstring(data)
+            microsoft_langs = [p.text for p in parsed.getchildren()]
+        except Exception, e:
+            logger.error('failed to get supported languages from Microsoft, using defaults (%s)', str(e))
+            microsoft_langs = ['ar','bg','ca','zh-CHS','zh-CHT','cs','da','nl','en','et','fi','fr','de','el','ht','he','hi','mww','hu','id','it','ja','ko','lv','lt','no','pl','pt','ro','ru','sk','sl','es','sv','th','tr','uk','vi']
+    else:
+        microsoft_langs = None
+
+    return render_to_response('config.js', RequestContext(request, {
+            'apertium_langs': apertium_langs,
+            'microsoft_langs': microsoft_langs,
+        }),
+        mimetype = 'application/javascript')